refactoring create account
